pyproject.toml
[tool.poetry]
name = "grader"
version = "0.1.0"
description = "CLI app to OCR handwritten tests and export cleaned per-student PDFs."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "grader" }]

[tool.poetry.dependencies]
python = "^3.11"
typer = { extras = ["all"], version = "^0.12.3" }
click = ">=8.1,<8.2"
rich = "^13.7.1"
pdf2image = "^1.17.0"
reportlab = "^4.0.8"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"
pydantic-settings = "^2.3.4"
regex = "^2024.4.16"
httpx = "^0.27.0"
anyio = "^4.3.0"
google-cloud-vision = "^3.7.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
black = "^24.4.2"
ruff = "^0.4.2"
types-requests = "^2.31.0.20240406"

[tool.poetry.scripts]
grader = "grader.cli:app"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "UP", "B", "C4"]

[tool.ruff.format]
quote-style = "double"
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
